# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema de Gerenciamento de Classificação Indicativa — Portaria 368/2014 MJ
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /usuarios:
    get:
      operationId: listUsuarios
      tags: [usuarios]
      summary: Listar usuários
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: createUsuario
      tags: [usuarios]
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios/{id}:
    get:
      operationId: getUsuario
      tags: [usuarios]
      summary: Obter usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
    patch:
      operationId: updateUsuario
      tags: [usuarios]
      summary: Atualizar usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
    delete:
      operationId: deleteUsuario
      tags: [usuarios]
      summary: Remover usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removido

  # ─── Conteúdos ────────────────────────────────────────────────────────────
...
  /solicitacoes:
    get:
      operationId: listSolicitacoes
      tags: [solicitacoes]
      summary: Listar solicitações de classificação
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: analista_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de solicitações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Solicitacao"
    post:
      operationId: createSolicitacao
      tags: [solicitacoes]
      summary: Abrir solicitação de classificação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitacaoInput"
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
  /solicitacoes/{id}:
    get:
      operationId: getSolicitacao
      tags: [solicitacoes]
      summary: Obter solicitação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Solicitação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
    patch:
      operationId: updateSolicitacao
      tags: [solicitacoes]
      summary: Atualizar solicitação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitacaoUpdate"
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
  /solicitacoes/{id}/aprovar:
    post:
      operationId: aprovarSolicitacao
      tags: [solicitacoes]
      summary: Aprovar e emitir classificação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AprovacaoInput"
      responses:
        "200":
          description: Aprovada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classificacao"
  /solicitacoes/{id}/rejeitar:
    post:
      operationId: rejeitarSolicitacao
      tags: [solicitacoes]
      summary: Rejeitar solicitação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejeicaoInput"
      responses:
        "200":
          description: Rejeitada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"

  # ─── Classificações ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Usuario:
      type: object
      required: [id, nome, email, perfil, ativo, criado_em]
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [administrador, supervisor, analista, solicitante]
        ativo:
          type: boolean
        criado_em:
          type: string
          format: date-time
    UsuarioInput:
      type: object
      required: [nome, email, perfil]
      properties:
        nome:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        perfil:
          type: string
          enum: [administrador, supervisor, analista, solicitante]
        senha:
          type: string
    UsuarioUpdate:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [administrador, supervisor, analista, solicitante]
        ativo:
          type: boolean

    # ─── Conteúdo ───────────────────────────────────────────────────────────
...
    Descritor:
      type: object
      required: [id, nome, categoria, descricao]
      properties:
        id:
          type: integer
        nome:
          type: string
        categoria:
          type: string
          enum: [violencia, sexualidade, drogas, linguagem, medo, outros]
        descricao:
          type: string
        icone:
          type: ["string", "null"]
...
    Solicitacao:
      type: object
      required: [id, conteudo_id, solicitante_id, status, criado_em]
      properties:
        id:
          type: integer
        conteudo_id:
          type: integer
        conteudo_titulo:
          type: ["string", "null"]
        conteudo_tipo:
          type: ["string", "null"]
        solicitante_id:
          type: integer
        solicitante_nome:
          type: ["string", "null"]
        analista_id:
          type: ["integer", "null"]
        analista_nome:
          type: ["string", "null"]
        status:
          type: string
          enum: [pendente, em_analise, aprovada, rejeitada]
        observacoes:
          type: ["string", "null"]
        prazo:
          type: ["string", "null"]
          format: date
        criado_em:
          type: string
          format: date-time
        atualizado_em:
          type: string
          format: date-time
    SolicitacaoInput:
      type: object
      required: [conteudo_id, solicitante_id]
      properties:
        conteudo_id:
          type: integer
        solicitante_id:
          type: integer
        observacoes:
          type: string
        prazo:
          type: string
          format: date
    SolicitacaoUpdate:
      type: object
      properties:
        analista_id:
          type: integer
        status:
          type: string
          enum: [pendente, em_analise, aprovada, rejeitada]
        observacoes:
          type: string
        prazo:
          type: string
          format: date

    # ─── Aprovação / Rejeição ───────────────────────────────────────────────
    AprovacaoInput:
      type: object
      required: [faixa_etaria, descritores_ids, analista_id]
      properties:
        faixa_etaria:
          type: string
          enum: [L, 10, 12, 14, 16, 18]
        descritores_ids:
          type: array
          items:
            type: integer
        analista_id:
          type: integer
        justificativa:
          type: string
        observacoes_publicas:
          type: string
        validade_anos:
          type: integer
    RejeicaoInput:
      type: object
      required: [motivo, analista_id]
      properties:
        motivo:
          type: string
          minLength: 10
        analista_id:
          type: integer

    # ─── Classificação ──────────────────────────────────────────────────────
    Classificacao:
      type: object
      required: [id, conteudo_id, solicitacao_id, faixa_etaria, analista_id, emitida_em]
      properties:
        id:
          type: integer
        conteudo_id:
          type: integer
        conteudo_titulo:
          type: ["string", "null"]
        conteudo_tipo:
          type: ["string", "null"]
        solicitacao_id:
          type: integer
        faixa_etaria:
          type: string
          enum: [L, 10, 12, 14, 16, 18]
        descritores:
          type: array
          items:
            $ref: "#/components/schemas/Descritor"
        analista_id:
          type: integer
        analista_nome:
          type: ["string", "null"]
        justificativa:
          type: ["string", "null"]
        observacoes_publicas:
          type: ["string", "null"]
        numero_certificado:
          type: string
        validade_anos:
          type: ["integer", "null"]
        emitida_em:
          type: string
          format: date-time
        atualizado_em:
          type: string
          format: date-time
...