# lib/api-zod/src/generated/types/aprovacaoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento de Classificação Indicativa — Portaria 368/2014 MJ
 * OpenAPI spec version: 0.1.0
 */
import type { AprovacaoInputFaixaEtaria } from './aprovacaoInputFaixaEtaria';

export interface AprovacaoInput {
  faixa_etaria: AprovacaoInputFaixaEtaria;
  descritores_ids: number[];
  analista_id: number;
  justificativa?: string;
  observacoes_publicas?: string;
  validade_anos?: number;
}