# lib/api-zod/src/generated/types/solicitacaoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento de Classificação Indicativa — Portaria 368/2014 MJ
 * OpenAPI spec version: 0.1.0
 */

export interface SolicitacaoInput {
  conteudo_id: number;
  solicitante_id: number;
  observacoes?: string;
  prazo?: Date;
}