# lib/api-zod/src/generated/types/solicitacaoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento de Classificação Indicativa — Portaria 368/2014 MJ
 * OpenAPI spec version: 0.1.0
 */
import type { SolicitacaoUpdateStatus } from './solicitacaoUpdateStatus';

export interface SolicitacaoUpdate {
  analista_id?: number;
  status?: SolicitacaoUpdateStatus;
  observacoes?: string;
  prazo?: Date;
}